Moved app to components
