+++ src/index.tsx
@@ -1,6 +1,6 @@
 import { render } from "solid-js/web";
 
-import App from "./App";
+import App from "./components/App.tsx";
 import "virtual:uno.css";
 import "./styles/tailwind-compat.css";
 import "./styles/icons.css"
